feat(transaction): refactor spending analytics into a dedicated screen
This commit extracts spending analytics from the transactions feature into a standalone screen. This refactor improves code separation by moving complex calculation logic out of the transactions ViewModel and providing a dedicated UI for spending trends and habit insights.

- **`AnalyticsViewModel.kt` & `AnalyticsUiRoute.kt`**:
    - Created a new ViewModel to handle the calculation of daily, weekly, and monthly spending trends.
    - Implemented `calculateHabitInsights` to identify frequent spending patterns and high-cost categories over the last 30 days.
    - Introduced a dedicated `AnalyticsUiRoute` featuring a `SimpleBarChart` for trends and a list of `HabitInsightCard` components.
- **`TransactionsUiRoute.kt` & `TransactionsViewModel.kt`**:
    - Removed the `Analytics` tab and its associated UI components (`AnalyticsPane`, `HabitInsightCard`).
    - Added an analytics icon to the `TopAppBar` to trigger navigation to the new dedicated screen.
    - Cleaned up the ViewModel and `TransactionsUiState` by removing analytics-related logic and state fields.
- **`SeaweedDestination.kt` & `seaweedNavEntryProvider.kt`**:
    - Defined a new `Analytics` destination in the navigation graph.
    - Removed `ANALYTICS` from the `TransactionTab` enum.
    - Registered the `AnalyticsUiRoute` in the main navigation entry provider.
- **`HabitInsight` Logic**:
    - Implemented logic to calculate monthly cost projections for frequent habits and identify high-frequency spending categories.
