+++ applications/seaweed/apps/mobile/features/transaction/src/main/java/com/zoewave/probase/seaweed/mobile/transaction/ui/TransactionsUiState.kt
@@ -1,8 +1,6 @@
 package com.zoewave.probase.seaweed.mobile.transaction.ui
 
 import com.zoewave.probase.seaweed.model.Transaction
-import com.zoewave.probase.seaweed.model.TrendPoint
-import com.zoewave.probase.seaweed.model.HabitInsight
 import com.zoewave.probase.seaweed.model.navigation.TransactionTab
 
 sealed interface TransactionsUiState {
@@ -14,9 +12,7 @@ sealed interface TransactionsUiState {
         val selectedCategory: String? = null,
         val selectedTransactionId: String? = null,
         val selectedTransaction: Transaction? = null,
-        val selectedTab: TransactionTab = TransactionTab.RECENT,
-        val spendingTrends: Map<com.zoewave.probase.seaweed.model.SpendingPeriod, List<TrendPoint>> = emptyMap(),
-        val habitInsights: List<HabitInsight> = emptyList()
+        val selectedTab: TransactionTab = TransactionTab.RECENT
     ) : TransactionsUiState
 }
